frontend/requirements.txt
nicegui>=1.4
httpx>=0.24
pydantic>=2.0
python-dotenv>=1.0
